requirements.txt
fastapi[standard]==0.115.12
pydantic==2.11.1
SQLAlchemy==2.0.40
sqlalchemy-utils==0.41.2
sqlmodel==0.0.24
psycopg==3.2.6
redis==5.2.1
uvicorn==0.34.0
pytest==8.3.5
pytest-mock==3.14
pytest-cov==6.1.0
black==25.1.0
isort==6.0.1
ipython==9.0.2
ipdb==0.13.13
polygon-api-client==1.14.4
lxml~=5.3.1
pgcli==4.3.0